Restructure project from clean  to all-in-one
