--- src/services/search-engine/Twitter.Clone.Search.Application/ConfigureServices.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-
-namespace Twitter.Clone.Search.Application;
-
-public static class ConfigureServices
-{
-    public static IServiceCollection RegisterApplicationServices(this IServiceCollection services)
-    {
-
-        return services;
-    }
-}
