--- src/services/search-engine/Twitter.Clone.Search.Infrastracture/ConfigureServices.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace Twitter.Clone.Search.Infrastracture;
-public static class ConfigureServices
-{
-    public static IServiceCollection RegisterInfrastructureServices(this IServiceCollection services)
-    {
-        services.AddDbContext<ApplicationDbContext>();
-
-        return services;
-    }
-}
