--- src/services/search-engine/Twitter.Clone.Search.Presentation/ConfigureServices.cs
@@ -1,15 +1,0 @@
-﻿namespace Twitter.Clone.Search.Presentation;
-
-public static class ConfigureServices
-{
-    public static IServiceCollection RegisterPresentationServices(this IServiceCollection services)
-    {
-
-        services.AddControllers();
-        services.AddEndpointsApiExplorer();
-        services.AddSwaggerGen();
-
-        return services;    
-    }
-
-}
