--- src/services/search-engine/Twitter.Clone.Search.Presentation/Twitter.Clone.Search.Presentation.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <InvariantGlobalization>true</InvariantGlobalization>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Twitter.Clone.Search.Application\Twitter.Clone.Search.Application.csproj" />
-    <ProjectReference Include="..\Twitter.Clone.Search.Infrastracture\Twitter.Clone.Search.Infrastracture.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Controllers\" />
-  </ItemGroup>
-
-</Project>
