--- src/services/search-engine/Twitter.Clone.Search.Presentation/Twitter.Clone.Search.Presentation.http
@@ -1,6 +1,0 @@
-@Twitter.Clone.Search.Presentation_HostAddress = http://localhost:5254
-
-GET {{Twitter.Clone.Search.Presentation_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
