+++ src/services/search-engine/Twitter.Clone.Search.WebApi/Data/ApplicationDbContext.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace Twitter.Clone.Search.Infrastracture;
+namespace Twitter.Clone.Search.WebApi.Data;
+
 public class ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : DbContext(options)
 {
+    
 }
