+++ src/services/search-engine/Twitter.Clone.Search.sln
@@ -3,23 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A4997026-5576-412C-8631-8432A3B2F37D}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{6F5CE995-B6F3-4283-88A8-A01E2EE63F74}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{E50C855A-43B6-479F-8C00-1D610C250BD4}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{07E46A3B-0F12-439E-9F10-DF9488DEE0D3}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentation", "Presentation", "{2B723D36-E0A5-4E1A-B23F-DDC798786205}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twitter.Clone.Search.Application", "Twitter.Clone.Search.Application\Twitter.Clone.Search.Application.csproj", "{9E4F3816-D341-4C70-873D-3EE86FFDC619}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twitter.Clone.Search.Domain", "Twitter.Clone.Search.Domain\Twitter.Clone.Search.Domain.csproj", "{79C8918F-C48C-489C-85A8-10C3DCE0499B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twitter.Clone.Search.Infrastracture", "Twitter.Clone.Search.Infrastracture\Twitter.Clone.Search.Infrastracture.csproj", "{0905B698-713D-46D8-8B18-B883A153D448}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twitter.Clone.Search.Presentation", "Twitter.Clone.Search.Presentation\Twitter.Clone.Search.Presentation.csproj", "{00C1EC91-80EA-4781-AAB5-49C543F9997F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Twitter.Clone.Search.WebApi", "Twitter.Clone.Search.WebApi\Twitter.Clone.Search.WebApi.csproj", "{9E7D402C-7440-4D4A-B69E-EEC4CCF28512}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,33 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9E4F3816-D341-4C70-873D-3EE86FFDC619}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9E4F3816-D341-4C70-873D-3EE86FFDC619}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9E4F3816-D341-4C70-873D-3EE86FFDC619}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9E4F3816-D341-4C70-873D-3EE86FFDC619}.Release|Any CPU.Build.0 = Release|Any CPU
-		{79C8918F-C48C-489C-85A8-10C3DCE0499B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79C8918F-C48C-489C-85A8-10C3DCE0499B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79C8918F-C48C-489C-85A8-10C3DCE0499B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79C8918F-C48C-489C-85A8-10C3DCE0499B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0905B698-713D-46D8-8B18-B883A153D448}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0905B698-713D-46D8-8B18-B883A153D448}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0905B698-713D-46D8-8B18-B883A153D448}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0905B698-713D-46D8-8B18-B883A153D448}.Release|Any CPU.Build.0 = Release|Any CPU
-		{00C1EC91-80EA-4781-AAB5-49C543F9997F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{00C1EC91-80EA-4781-AAB5-49C543F9997F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{00C1EC91-80EA-4781-AAB5-49C543F9997F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{00C1EC91-80EA-4781-AAB5-49C543F9997F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9E7D402C-7440-4D4A-B69E-EEC4CCF28512}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E7D402C-7440-4D4A-B69E-EEC4CCF28512}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E7D402C-7440-4D4A-B69E-EEC4CCF28512}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E7D402C-7440-4D4A-B69E-EEC4CCF28512}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{E50C855A-43B6-479F-8C00-1D610C250BD4} = {A4997026-5576-412C-8631-8432A3B2F37D}
-		{07E46A3B-0F12-439E-9F10-DF9488DEE0D3} = {A4997026-5576-412C-8631-8432A3B2F37D}
-		{2B723D36-E0A5-4E1A-B23F-DDC798786205} = {A4997026-5576-412C-8631-8432A3B2F37D}
-		{9E4F3816-D341-4C70-873D-3EE86FFDC619} = {E50C855A-43B6-479F-8C00-1D610C250BD4}
-		{79C8918F-C48C-489C-85A8-10C3DCE0499B} = {E50C855A-43B6-479F-8C00-1D610C250BD4}
-		{0905B698-713D-46D8-8B18-B883A153D448} = {07E46A3B-0F12-439E-9F10-DF9488DEE0D3}
-		{00C1EC91-80EA-4781-AAB5-49C543F9997F} = {2B723D36-E0A5-4E1A-B23F-DDC798786205}
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7AE024D9-B15B-42F9-8253-64D743B258BA}
 	EndGlobalSection
 EndGlobal
